Finished authorization policies

+++ app/controllers/collabs_controller.rb
@@ -1,11 +1,12 @@
 class CollabsController < ApplicationController
   before_action :authenticate_user!
   after_action :verify_authorized, except: :index
-  after_action :verify_policy_scoped, only: :index
+  # after_action :verify_policy_scoped, only: :index
 
 
   def index
     @wiki = Wiki.find(params[:wiki_id])
+    authorize @wiki, :show_collabs?
     @owner = User.owner_by_wiki(@wiki.id)
     @current = User.current_collaborators(@wiki.id)
     @available = User.where.not(id: @owner.id) - @current.to_a
@@ -13,7 +14,9 @@ class CollabsController < ApplicationController
 
   def create
     @wiki = Wiki.find(params[:wiki_id])
+    authorize @wiki, :collab_add_remove?
     @collab = Collab.new(collab_params)
+    authorize @collab
 
     if @collab.save
       redirect_to wiki_collabs_path, notice: 'Collaborator successfully added.'
@@ -23,6 +26,9 @@ class CollabsController < ApplicationController
   end
 
   def destroy
+    @wiki = Wiki.find(params[:wiki_id])
+    authorize @wiki, :collab_add_remove?
+
     @collab = Collab.find(params[:id])
 
     if @collab.destroy
